server/requirements.txt
Flask==2.2.5
flask-login==0.6.2
flask-sqlalchemy==3.0.5
psycopg2-binary==2.9.9
python-dotenv>=0.19.2
Werkzeug==2.3.7
PyJWT==2.1.0
email-validator==1.1.3
Flask-Mail==0.9.1
oauthlib==3.2.2
requests-oauthlib==1.3.1
flask-wtf==1.2.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
alembic==1.12.1
scikit-learn>=1.0.2
requests>=2.27.1
python-dateutil>=2.8.2
gunicorn==21.2.0
Flask-Session==0.5.0
redis>=4.0.0